python/agents/order-processing/pyproject.toml
[project]
name = "order-processing"
version = "0.1.0"
description = "AI agent that processes orders and orchestrates human in the loop pattern if the order quantity exceeds 100"
authors = [
    {name = "Sai Saran Vaidyanathan", email = "ssvaidyanathan@google.com"},
    {name = "Matt Hartmann", email = "hartmannm@google.com"}
]
license = {text = "Apache License 2.0"}
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "google-adk>=1.5.0,<2.0.0",
    "google-cloud-aiplatform[adk,agent-engines]>=1.100.0,<2.0.0",
    "python-dotenv>=1.1.1,<2.0.0",
    "google-cloud-secret-manager>=2.24.0,<3.0.0",
]

[dependency-groups]
dev = [
    "agent-starter-pack>=0.32.0",
    "pytest>=8.0.0",
    # google-adk is already in main dependencies
]
deployment = [
    "absl-py>=2.2.2",
]

[tool.ruff]
extend = "../../../pyproject.toml"

[tool.ruff.lint.isort]
known-first-party = ["order_processing"]

[tool.pytest.ini_options]
pythonpath = "."


[tool.agent-starter-pack]
example_question = "I'd like to order 5 Google Pixel phones. My name is Test User and my address is 15 Testing Drive."

[tool.agent-starter-pack.settings]
agent_directory = "order_processing"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
